refactor(yaml): modularize ds/dd; kinds as primkey
