+++ type/zkgyaml.js
@@ -1,159 +1,15 @@
 import yaml from "js-yaml";
 import fs from "fs";
-import { ethers } from "ethers";
 import semver from "semver";
 import { YamlHealthCheckFailed } from "../common/error.js";
+import { EthereumDataDestination, EthereumDataSource } from "./zkgyaml_eth.js";
+import { Mapping } from "./zkgyaml_def.js";
 
-function isEthereumAddress(address) {
-  try {
-    const parsedAddress = ethers.utils.getAddress(address);
-    return parsedAddress !== '0x0000000000000000000000000000000000000000';
-  } catch (error) {
-    return false;
-  }
-}
+const dataSourceClassMap = new Map();
+dataSourceClassMap.set('ethereum', EthereumDataSource)
 
-class DataSource {
-  constructor(kind) {
-    this.kind = kind;
-  }
-}
-
-class DataDestination {
-  constructor(kind) {
-    this.kind = kind;
-  }
-}
-
-class Mapping {
-  constructor(language, file, handler) {
-    this.language = language;
-    this.file = file;
-    this.handler = handler;
-  }
-  static from_v_0_0_2(yamlMapping){
-    return new Mapping(
-      yamlMapping.language,
-      yamlMapping.file,
-      yamlMapping.handler
-    )
-  }
-  static from_v_0_0_1(yamlMapping){
-    return null // not important for 0.0.1
-  }
-}
-
-class EthereumDataSource extends DataSource {
-  constructor(kind, network, eventSection, storageSection, accountSection, blockSection) {
-    super(kind);
-    this.network = network;
-    this.event = eventSection;
-    this.storage = storageSection;
-    this.account = accountSection;
-    this.block = blockSection;
-  }
-
-  static from_v_0_0_2(yamlEthDS){
-    return new EthereumDataSource(
-      yamlEthDS.kind,
-      yamlEthDS.network,
-      (yamlEthDS.event != null) ?  EventSection.from_v_0_0_2(yamlEthDS.event) : null,
-      (yamlEthDS.storage != null) ? StorageSection.from_v_0_0_2(yamlEthDS.storage) : null,
-      null, //TODO: account section
-      null // TODO: block section
-    )
-  }
-  static from_v_0_0_1(yamlEthDS){
-    throw new Error('no 0.0.1 support'); //TODO
-  }
-}
-
-class EthereumDataDestination extends DataDestination{
-  constructor(kind, network, address) {
-    super(kind);
-    this.network = network;
-    this.address = address;
-  }
-  static from_v_0_0_2(yamlEthDD){
-    return new EthereumDataDestination(
-      yamlEthDD.kind,
-      yamlEthDD.network,
-      yamlEthDD.address
-    )
-  }
-
-  static from_v_0_0_1(yamlEthDD){
-    return new EthereumDataDestination(
-      yamlEthDD.kind,
-      yamlEthDD.network,
-      yamlEthDD.destination.address
-    )
-  }
-}
-
-class EventSection {
-  constructor(addressList, esigsList) {
-    this.addressList = addressList;
-    this.esigsList = esigsList;
-  }
-
-  static from_v_0_0_2(yamlEvent){
-    let loadFromEventSource = (event) => {
-      const source_address = event.address;
-        const source_esigs = event.events.map((ed) => {
-          const eventHash = ed.startsWith("0x") ? ed : ethers.utils.keccak256(ethers.utils.toUtf8Bytes(ed));
-          return eventHash;
-        });
-  
-        return [source_address, source_esigs];
-    }
-  
-    const eventDSAddrList=[];
-    const eventDSEsigsList=[];
-    yamlEvent.map((event) => {let [sa, se] = loadFromEventSource(event); eventDSAddrList.push(sa); eventDSEsigsList.push(se)})
-    
-    return new EventSection(eventDSAddrList, eventDSEsigsList)
-  }
-
-  static from_v_0_0_1(yamlEvent){
-    throw new Error('no 0.0.1 support'); //TODO
-  }
-  
-  toArray(){
-    return [this.addressList, this.esigsList]
-  }
-}
-
-class StorageSection {
-  constructor(addressList, esigsList) {
-    this.addressList = addressList;
-    this.slotsList = esigsList;
-  }
-
-  static from_v_0_0_2(yamlStorage){
-    let loadFromStorageSource = (storage) => {
-      const source_address = storage.address;
-        const source_slots = storage.slots.map((sl) => {
-          return ethers.utils.hexZeroPad(sl, 32);;
-        });
-
-        return [source_address, source_slots];
-    }
-
-    const stateDSAddrList=[];
-    const stateDSSlotsList=[];
-    yamlStorage.map((storage) => {let [sa, sl] = loadFromStorageSource(storage); stateDSAddrList.push(sa); stateDSSlotsList.push(sl)})
-    return new StorageSection(stateDSAddrList, stateDSSlotsList);
-  }
-
-  static from_v_0_0_1(yamlStorage){
-    throw new Error('no 0.0.1 support'); //TODO
-  }
-
-  toArray(){
-    return [this.addressList, this.slotsList]
-  }
-}
+const dataDestinationClassMap = new Map();
+dataDestinationClassMap.set('ethereum', EthereumDataDestination)
 
 export class ZkGraphYaml {
   constructor(specVersion, apiVersion, name, description, repository, dataSources, dataDestinations, mapping) {
@@ -168,9 +24,9 @@ export class ZkGraphYaml {
   }
   static from_v_0_0_2(yaml) {
     const dataSources = [];
-    yaml.dataSources.forEach(ds => dataSources.push(EthereumDataSource.from_v_0_0_2(ds)));
+    yaml.dataSources.forEach(ds => dataSources.push(dataSourceClassMap.get(ds.kind).from_v_0_0_2(ds)));
     const dataDestinations = [];
-    yaml.dataDestinations.forEach(dd => dataDestinations.push(EthereumDataDestination.from_v_0_0_2(dd)));
+    yaml.dataDestinations.forEach(dd => dataDestinations.push(dataDestinationClassMap.get(dd.kind).from_v_0_0_2(dd)));
 
     return new ZkGraphYaml(
       yaml.specVersion,
@@ -220,21 +76,29 @@ export class ZkGraphYaml {
     throw new Error('no 0.0.1 support'); //TODO
   }
   
-  // const config = ;
-  sourceType() { //TODO
-    if (this.specVersion >= "0.0.2"){
-      if (this.dataSources[0].event) {
-        return "event";
-      } else if (this.dataSources[0].storage) {
-        return "storage";
-      } else {
-        throw new Error("At least includes event or storage section in dataSources[0]")
-      }
-    } else if (this.specVersion == "0.0.1"){
-      return "event"; // 0.0.1 only support event
-    }
+  // // const config = ;
+  // sourceType() { //TODO
+  //   if (this.specVersion >= "0.0.2"){
+  //     if (this.dataSources[0].event) {
+  //       return "event";
+  //     } else if (this.dataSources[0].storage) {
+  //       return "storage";
+  //     } else {
+  //       throw new Error("At least includes event or storage section in dataSources[0]")
+  //     }
+  //   } else if (this.specVersion == "0.0.1"){
+  //     return "event"; // 0.0.1 only support event
+  //   }
 
-    throw new Error("Unsupported specVersion: ", this.specVersion)
+  //   throw new Error("Unsupported specVersion: ", this.specVersion)
+  // }
+
+  getKinds(isSource){
+    return isSource ? this.dataSources.map(ds => ds.kind) : this.dataDestinations.map(ds => ds.kind)
+  }
+
+  getFilteredSourcesByKind(kind){
+    return this.dataSources.filter(ds => ds.kind === kind)
   }
 
   healthCheck() {
@@ -262,27 +126,28 @@ export class ZkGraphYaml {
       throw new YamlHealthCheckFailed("dataSources should not be empty");
     }
 
-    const sourceNetworks = [];
+    const sourceKinds = [];
 
     this.dataSources.forEach(dataSource => {
-      // every object in datasources MUST have network
-      if (!dataSource.kind || !dataSource.network) {
-        throw new YamlHealthCheckFailed("dataSource object is missing required fields");
+      // every object in datasources MUST have kind
+      if (!dataSource.kind) {
+        throw new YamlHealthCheckFailed("dataSource is missing 'kind' field");
       }
 
-      sourceNetworks.push(dataSource.network);
+     // check data sources
+      dataSourceClassMap.get(dataSource.kind).healthCheck(dataSource);
 
-      const eventCount = dataSource.event ? 1 : 0;
-      const storageCount = dataSource.storage ? 1 : 0;
-
-      if (eventCount + storageCount !== 1) {
-        throw new YamlHealthCheckFailed("must have one and only one 'event' or 'storage' field");
-      }
+      sourceKinds.push(dataSource.kind);
     });
 
     // every network field must be the same
-    if (new Set(sourceNetworks).size !== 1) {
-      throw new YamlHealthCheckFailed("All dataSource networks must be the same");
+    // if (new Set(sourceKinds).size !== 1) {
+    //   throw new YamlHealthCheckFailed("All dataSource networks must be the same");
+    // }
+
+    // can only have 1 data source with kind 'ethereum' right now
+    if (sourceKinds.indexOf('ethereum') != sourceKinds.lastIndexOf('ethereum')){
+      throw new YamlHealthCheckFailed("Only 1 'ethereum' kind is allowed in data sources right now");
     }
 
     // all mapping fields must be not empty
@@ -290,17 +155,14 @@ export class ZkGraphYaml {
       throw new YamlHealthCheckFailed("Some required fields are empty in mapping");
     }
 
-    // data destination must have network and destination
-    if (this.dataDestinations) {
-      if (!this.dataDestinations[0].network || !this.dataDestinations[0].address) {
-        throw new YamlHealthCheckFailed("dataDestinations object is missing required fields");
+    this.dataDestinations.forEach(dataDest => {
+      // every object in datasources MUST have kind
+      if (!dataDest.kind) {
+        throw new YamlHealthCheckFailed("dataDestination is missing 'kind' field");
       }
-
-      // address must be the ethereum address and not address zero
-      if (!isEthereumAddress(this.dataDestinations[0].address)) {
-        throw new YamlHealthCheckFailed("Invalid Ethereum address in dataDestinations");
-      }
-    }
+      // check data destinations
+      dataDestinationClassMap.get(dataDest.kind).healthCheck(dataDest);
+    });
 
     // 12. the network must be same as the source network
     // TODO: right now we don't check the block hash, so skip the same network check
